+++ sources/ewol/widget/meta/ColorChooser.h
@@ -45,10 +45,10 @@ namespace ewol {
 				std::shared_ptr<ewol::widget::Slider> m_widgetBlue;
 				std::shared_ptr<ewol::widget::Slider> m_widgetAlpha;
 				etk::Color<> m_currentColor;
-				void onCallbackColorChangeRed(const int32_t& _newColor);
-				void onCallbackColorChangeGreen(const int32_t& _newColor);
-				void onCallbackColorChangeBlue(const int32_t& _newColor);
-				void onCallbackColorChangeAlpha(const int32_t& _newColor);
+				void onCallbackColorChangeRed(const float& _newColor);
+				void onCallbackColorChangeGreen(const float& _newColor);
+				void onCallbackColorChangeBlue(const float& _newColor);
+				void onCallbackColorChangeAlpha(const float& _newColor);
 				void onCallbackColorChange(const etk::Color<>& _newColor);
 		};
 	};
